feat: Add linting with ESLint
ESLint is a tool that formats codebases, no matter whose behind the
keyboard. With Prettier and Pretty-Quick, any modified files that have
changed versus the master branch will be formatted accordingly. This
will ensure that any files accidentally touched are returned to their
former state, resulting in PR's that are less polluted.

Fixes #9